refactor: move middlewares to drivers
