+++ src/app/article/api.rs
@@ -4,8 +4,8 @@ use super::{
     response::{MultipleArticlesResponse, SingleArticleResponse},
     service,
 };
-use crate::middleware::auth;
-use crate::middleware::state::AppState;
+use crate::appv2::drivers::middlewares::auth;
+use crate::appv2::drivers::middlewares::state::AppState;
 use crate::utils::api::ApiResponse;
 use actix_web::{web, HttpRequest, HttpResponse};
 use serde::Deserialize;
