+++ src/app/tag/api.rs
@@ -1,7 +1,7 @@
 extern crate serde_json;
 use super::model::Tag;
 use super::response::TagsResponse;
-use crate::middleware::state::AppState;
+use crate::appv2::drivers::middlewares::state::AppState;
 use crate::utils::api::ApiResponse;
 use actix_web::{web, HttpResponse};
