+++ src/app/user/api.rs
@@ -1,7 +1,7 @@
 use super::model::{UpdateUser, User};
 use super::{request, response::UserResponse};
-use crate::middleware::auth;
-use crate::middleware::state::AppState;
+use crate::appv2::drivers::middlewares::auth;
+use crate::appv2::drivers::middlewares::state::AppState;
 use crate::utils::api::ApiResponse;
 use actix_web::{web, HttpRequest, HttpResponse};
