+++ src/appv2/drivers/mod.rs
@@ -1,1 +1,2 @@
+pub mod middlewares;
 pub mod routes;
